ValidacionesEntity

+++ src/main/java/com/ipn/mx/domain/Controller/ArticuloController.java
@@ -3,8 +3,11 @@ package com.ipn.mx.domain.Controller;
 
 import com.ipn.mx.domain.Entity.Articulo;
 import com.ipn.mx.domain.Entity.Categoria;
+import com.ipn.mx.domain.Repository.ArticuloRepository;
 import com.ipn.mx.domain.Service.ArticuloService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,19 +18,27 @@ import java.util.List;
 @RequestMapping("/articulo")
 public class ArticuloController{
     private final ArticuloService articuloService;
+    private final ArticuloRepository articuloRepository;
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public void createArticulo(@RequestBody Articulo articulo) {
         // Create a new articulo
         articuloService.createArticulo(articulo);
     }
+
     @GetMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
     public Articulo getArticulo(@PathVariable Long id) {
         // Get articulo by id
+        if (id<=0 || articuloRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Articulo no encontrado");
+        }
         return articuloService.getArticulo(id);
     }
 
     @PutMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
     public void updateArticulo(@PathVariable Long id, @RequestBody Articulo articulo) {
         // Update an existing articulo
         if (!id.equals(articulo.getIdArticulo())) {
@@ -37,17 +48,23 @@ public class ArticuloController{
     }
 
     @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteArticulo(@PathVariable Long id) {
+        if (id<=0 || articuloRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Articulo no encontrado");
+        }
         // Delete an existing articulo
         articuloService.deleteArticulo(id);
     }
 
     @GetMapping
-    public List<Articulo> getAllArticulos(){
+    @ResponseStatus(HttpStatus.OK)
+    public List<Articulo> getAllArticulos() {
         return articuloService.getAllArticulos();
     }
 
     @GetMapping("/categoria/{idCategoria}")
+    @ResponseStatus(HttpStatus.OK)
     public List<Articulo> getArticulosByCategoria(@PathVariable Long idCategoria) {
         // Get all articulos by categoria
         return articuloService.getArticulosByCategoria(idCategoria);
